+++ CHANGELOG.md
@@ -1,3 +1,25 @@
+## 0.0.8
+- fixed mistake in readme
+- added `body` field to the `delete` methods
+## 0.0.7
+- Renamed `getSingle` to `getOne` and `getMaybeSingle` to `getOneOrNull`, this should be more in line
+with general naming conventions
+
+## 0.0.6
+- Renamed `baseExpansions` to just `expansions` on the `CollectionHelper` class
+- Added a getter for the raw `RecordService` object to `CollectionHelper`
+- added a new getMaybeSingle() method to `CollectionHelper` and `BaseHelper`
+- renamed `otherFilters` to `additionalExpressions` on search methods,
+- renamed `otherParams` to `additionalParams` on search methods
+- removed `SearchParams` class and put its fields directly into the search method.
+This was done to make the method more intuitive to use
+- added `fields` parameter to all methods, takes a list of fields to be included in the returning record
+if not given all fields are returned
+- Fixed an issue where expansions would not propegate in `addFiles` and `removeFiles` methods
+- added queryParameters optionalfields to the getFileUrl method
+- renamed `creationHook` and `updateHook` to `preCreationHook` and `preUpdateHook`
+- updated readme and inline docs
+
 ## 0.0.5
 - Fixed breaking bug with filtering in fullList fetch